Passage des double en float

+++ serveur/src/exoTaux/session/TauxItf.java
@@ -7,8 +7,8 @@ import java.util.Collection;
 @Remote
 public interface TauxItf {
 
-    public double conversion(String uneMonnaieA, String uneMonnaieB, double unMontant) throws Exception;
-    public void changer(String uneMonnaieA, String uneMonnaieB, double nouveauTaux) throws Exception;
+    public float conversion(String uneMonnaieA, String uneMonnaieB, float unMontant) throws Exception;
+    public void changer(String uneMonnaieA, String uneMonnaieB, float nouveauTaux) throws Exception;
     public Collection<String> getAllTitle();
 
 }
